face_rec/requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.3
dlib @ file:///C:/Users/Saif%20Pc/Desktop/face_rec/dlib-19.22.99-cp310-cp310-win_amd64.whl
easydict==1.13
exceptiongroup==1.3.1
faiss-cpu==1.7.4
fastapi==0.128.0
flatbuffers==25.12.19
fonttools==4.61.1
h11==0.16.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface @ file:///C:/Users/Saif%20Pc/Desktop/face_rec/insightface-0.7.3-cp310-cp310-win_amd64.whl
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
mediapipe==0.10.9
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
onnx==1.20.0
onnxruntime==1.15.1
opencv-contrib-python==4.12.0.88
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.21
PyYAML==6.0.3
qudida==0.0.4
requests==2.32.5
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.10.1
simsimd==6.5.12
six==1.17.0
sounddevice==0.5.3
starlette==0.50.0
stringzilla==4.5.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
